pyproject.toml
[project]
name = "hospitalpipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "Nicolas",email = "nicolascnagel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "azure-identity (>=1.25.1,<2.0.0)",
    "azure-storage-blob (>=12.28.0,<13.0.0)",
    "pandera (>=0.28.1,<0.29.0)",
    "pyarrow (>=22.0.0,<23.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
